backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.125.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.13.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pyyaml==6.0.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
alembic==1.13.1
python-multipart==0.0.21
pandoc              # 文档转换
python-docx         # Word 文件处理
openpyxl            # Excel 文件处理
PyPDF2              # PDF 文件处理
aiomysql
greenlet
